feat: add TypeORM module and configure data source
This commit adds the TypeORM module to the app module and configures the data source using the provided options. The data source options are defined in a new file called `data-source.ts`. This allows the application to connect to a PostgreSQL database with the specified credentials and database name. The `synchronize` and `logging` options are set to `false` to disable automatic schema synchronization and logging.